requirements.txt
groq
python-dotenv
SpeechRecognition
pyaudio
pyttsx3
pywin32
duckduckgo-search

